refactor: utilities folder to modules.
