+++ src/common/modules/mailer/models/customer-staking-request-service.model.ts
