+++ src/common/modules/mailer/templates/customer-staking-request-service.hbs
