+++ src/common/modules/mailer/templates/lab-register.hbs
