+++ src/common/modules/reward/models/reward.entity.ts
