+++ src/endpoints/ethereum/ethereum.service.ts
@@ -9,11 +9,10 @@ import {
   getFile,
   createFile,
   checkIfFileOrDirectoryExists,
-} from '../../common/utilities/helper/storage.helper';
+} from '../../common/modules/helper/storage.helper';
 import ABI from './utils/ABI.json';
 import axios from 'axios';
 import escrowContract from './utils/Escrow.json';
-import serviceRequestContract from './utils/ServiceRequest.json';
 import { ethers } from 'ethers';
 
 @Injectable()
